Added support for various modes of reporting.
Pakiti server can be instructed by the client to process the report in one
of three modes: store-only, report-only, store-and-report. The indication is
passed in the query string in variable mode, store-only is used by default.

The naming was chosen to avoid confusion about various report types. Note that
Pakiti2 client used another name for the protocol option ('type') with values
being integeres not constant strings.