Handle protocol identifier properly.
The server supports two protocol versions:
version 4 == format used by legacy Pakiti2 clients
version 5 == format introduced to support wLCG middleware reporter

The client sends the protocol version as part of the message envelope, i.e.
in the URL query string not in the report. Version 5 is expected when the
protocol version isn't specified by the client.

The handling of the version was made a bit more robust on the server.

The client was changed to always include the protocol version (5) in the
messages reported to server.