Extended functions of the hosts.php CLI tool.
The command now makes it possible to specify the activity parameter to limit
the list of displayed hosts. The output now include also the time of the last
report. In order to ease the use with multi-instance environments, the command
accepts a --config parameter to point to the configuration file.