Add support for CentOS versioning.
In some cases CentOS selects to add own changes to a package, which yields a
new version number of the package. The new version can't be matched against
RH-issued vulnerability information, which makes custom CentOS packages
unverifiable.

Since CentOS doesn't publishe a reasonable source on fixed vulnerabilities,
this patch adds a possibility to craft CentOS-specific infomation, based on
the RH ones. It's quite inefficient and still might not be 100% covering the
space.