Use utf8 as the default encoding for Pakiti DB.
The utf8 encoding requires less space then the default utf8m4, which is used on
deb9. When utf8m4 was used some tables couldn't be created since they would
need more space then available and the process failed with an error. The
encoding change makes it possible to keep the current schema without additional
changes to Mysql configuration. We should always use utf-8 (if not ascii :-),
so the change should be harmless.