Improve handling of kernel packages on storing.
Add a fake kernel record for deployments that don't have the actual kernel
package installed, which makes it possible to check the running kernel version.
Add a fake kernel record for Debian systems named 'linux', which is necessary
to match the packege names against vulnerability definitions (which are based
on source package names).

The latter change would require a more systematic solution as the problem can
appear for other packages, too!