Move processing of vulnerability definitions to subsources
The SubSource class is now responsible for entire processing of vulnerability
definitions that are read from external sources. The change renders the Source
class quite useless, which could simplify the code base. A side-effect of the
change is better control the lastAccess times for source definitions.

The patch moves a great deal of code, ignore white-space when examining the
changes.