Support properly Debian advisories
Debian refers to source packages in their security advisories, and the mapping
to the corresponding binary packages (that are present on the end system) has
to be established. This change relyies on the Source indices that are
downloaded on demand in the course of parsing the primary advisory source.

This patch add two configuration directives needed to control the flow.